init configService && GlobalConfigService
