+++ src/base/browser/secondary/contextMenu/editor/editorContextMenu.ts
@@ -1,5 +1,4 @@
 import { ContextMenu, ContextMenuType, Coordinate, IContextMenu } from "src/base/browser/secondary/contextMenu/contextMenu";
-import { ipcRendererSend } from "src/base/electron/register";
 import { IComponentService } from "src/code/browser/service/componentService";
 import { IContextMenuService } from "src/code/browser/service/contextMenuService";
 
@@ -46,7 +45,7 @@ export class EditorContextMenu extends ContextMenu implements IContextMenu {
 
         document.getElementById('select all')!.addEventListener('click', (ev) => {
             document.execCommand("selectAll");
-            this.contextMenuService..removeContextMenu();
+            this.contextMenuService.removeContextMenu();
         });
 
     }
