+++ src/code/browser/workbench/editor/titleBar/functionBar.ts
@@ -1,7 +1,7 @@
 import { Button } from "src/base/browser/basic/button";
 import { EVENT_EMITTER } from "src/base/common/event";
 import { getSvgPathByName, SvgType } from "src/base/common/string";
-import { ConfigModule } from "src/base/config";
+import { ConfigService } from "src/code/common/service/configService";
 import { domNodeByIdAddListener } from "src/base/electron/register";
 import { IComponentService } from "src/code/browser/service/componentService";
 import { Component } from "src/code/browser/workbench/component";
@@ -62,7 +62,7 @@ export class FunctionBarComponent extends Component {
      * @description function calls when the functionBarModule is initialized.
      */
      initfunctionBar(): void {
-        if (ConfigModule.Instance.defaultMarkdownMode == 'wysiwyg') {
+        if (ConfigService.Instance.defaultMarkdownMode == 'wysiwyg') {
             $('#mode-switch').addClass('function-button-focus');
         }
