+++ src/code/browser/workbench/workbench.ts
@@ -5,12 +5,13 @@ import { EditorComponent, IEditorService } from "src/code/browser/workbench/edit
 import { LOCAL_MDNOTE_DIR_NAME, NoteBookManager } from "src/code/common/model/notebookManger";
 import { APP_ROOT_PATH } from "src/base/electron/app";
 import { ipcRendererOn, ipcRendererSend } from "src/base/electron/register";
-import { ConfigModule, DEFAULT_CONFIG_FILE_NAME, DEFAULT_CONFIG_PATH, GlobalConfigModule, GLOBAL_CONFIG_FILE_NAME, GLOBAL_CONFIG_PATH, LOCAL_CONFIG_FILE_NAME } from "src/base/config";
+import { ConfigService, DEFAULT_CONFIG_FILE_NAME, DEFAULT_CONFIG_PATH, GLOBAL_CONFIG_FILE_NAME, GLOBAL_CONFIG_PATH, LOCAL_CONFIG_FILE_NAME } from "src/code/common/service/configService";
 import { pathJoin } from "src/base/common/string";
 import { ContextMenuService, IContextMenuService } from 'src/code/browser/service/contextMenuService';
 import { IInstantiationService } from "src/code/common/service/instantiation/instantiation";
 import { ServiceDescriptor } from "src/code/common/service/instantiation/descriptor";
 import { ComponentService } from "src/code/browser/service/componentService";
+import { GlobalConfigService } from "src/code/common/service/globalConfigService";
 
 /**
  * @description Workbench represents all the Components in the web browser.
@@ -49,8 +50,6 @@ export class Workbench extends Component {
         // ContextMenuService
         this.instantiationService.register(IContextMenuService, new ServiceDescriptor(ContextMenuService));
 
-        // ComponentService
-
     }
 
     /**
@@ -84,17 +83,17 @@ export class Workbench extends Component {
         ipcRendererOn('closingApp', () => {
             
             // save global configuration first
-            GlobalConfigModule.Instance.previousNoteBookManagerDir = this._noteBookManager.noteBookManagerRootPath;
-            GlobalConfigModule.Instance.writeToJSON(GLOBAL_CONFIG_PATH, GLOBAL_CONFIG_FILE_NAME)
+            GlobalConfigService.Instance.previousNoteBookManagerDir = this._noteBookManager.noteBookManagerRootPath;
+            GlobalConfigService.Instance.writeToJSON(GLOBAL_CONFIG_PATH, GLOBAL_CONFIG_FILE_NAME)
             .then(() => {
                 // save local or default configuration
-                if (GlobalConfigModule.Instance.defaultConfigOn) {
-                    return ConfigModule.Instance.writeToJSON(
+                if (GlobalConfigService.Instance.defaultConfigOn) {
+                    return ConfigService.Instance.writeToJSON(
                         DEFAULT_CONFIG_PATH, 
                         DEFAULT_CONFIG_FILE_NAME
                     );
                 }
-                return ConfigModule.Instance.writeToJSON(
+                return ConfigService.Instance.writeToJSON(
                     pathJoin(this._noteBookManager.noteBookManagerRootPath, LOCAL_MDNOTE_DIR_NAME), 
                     LOCAL_CONFIG_FILE_NAME
                 );
